refactor: move Card Menu reveal logic to Card component
