+++ app/components/invitation/card.html.haml
@@ -8,7 +8,7 @@
     .contents{ stimulus_target_option("pageAttachment", self.stimulus_identifier) }
       -# Page Navigation
       - options = ::Html::TagAttributes.build({ class: "absolute flex flex-row items-center gap-6 top-[-90px] right-8 z-[-1]" })
-      - options = options.with_stimulus_target("menu", Invitation::Envelop.stimulus_identifier)
+      - options = options.with_stimulus_target("menu", stimulus_identifier)
       %div{ options.to_h }
         - options = ::Html::TagAttributes.build({ class: "opacity-0" }).with_stimulus_action("click", Invitation::Card.stimulus_identifier, "navigateTo")
