+++ app/components/invitation/envelop.rb
@@ -12,6 +12,11 @@ class Invitation::Envelop < Invitation::ApplicationComponent
 
   def default_content_tag_arguments
     options = ::Html::TagAttributes.build(super, { class: "w-fit", style: "perspective: 1600px" })
+    options = options.with_stimulus_action(
+      "#{Invitation::Card.stimulus_identifier}:connected",
+      stimulus_identifier,
+      "registerCardController"
+    )
     options = options.with_stimulus_controller(self.class.stimulus_identifier)
     options.to_h
   end
